Add solution for problem `D - Make The Fence Great Again` Link: https://codeforces.com/contest/1221/problem/D Tags: dp, *1800 Ref: https://codeforces.com/contest/1221/submission/70167020

+++ codeforces/1221/D.cpp
@@ -155,7 +155,7 @@ signed main(int argc, char** argv)
                 {
                     if(a[i]+j!=a[i-1]+k)
                     {
-                        temp = min(temp, tab[k][i-1]+b[i]*j);
+                        temp = min(temp, tab[k][i-1]+b[i]*(ll)j);
                     }
                 }
                 tab[j][i]=temp;
